Add Windows LTSC version to label

+++ runner/configure.ps1
@@ -9,8 +9,13 @@ if (!($env:REPO)) {
   Write-Error 'Variable "TOKEN" is not set.'
   exit 2
 }
+if (!($env:LTSC_YEAR)) {
+  Write-Error 'Variable "LTSC_YEAR" is not set.'
+  exit 2
+}
 $TOKEN = $env:TOKEN
 $REPO = $env:REPO
+$LTSC_YEAR = $env:LTSC_YEAR
 $GROUP = $env:GROUP
 if ($env:GROUP) {
   Remove-Item Env:\GROUP
@@ -19,7 +24,8 @@ if ($env:GROUP) {
 }
 Remove-Item Env:\TOKEN
 Remove-Item Env:\REPO
-. .\config.cmd --unattended --url https://github.com/$REPO --token $TOKEN --runnergroup $GROUP
+Remove-Item Env:\LTSC_YEAR
+. .\config.cmd --unattended --url https://github.com/$REPO --token $TOKEN --runnergroup $GROUP --labels windows-$LTSC_YEAR
 if ($LASTEXITCODE -ne 0) {
   Write-Error "Failed to configure GitHub Actions runner."
   exit 1
